refactor: modularize converter
