+++ index.html
@@ -128,6 +128,6 @@
             <a href="sitemap.html">Plan du site</a>
         </nav>
     </footer>
-    <script src="script.js"></script>
+    <script type="module" src="js/main.js"></script>
 </body>
 </html>
